requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.6
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
tweepy==4.14.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
